Seperated router files and added enums for github actions.
