+++ src/lib/enums.lib.ts
@@ -1,0 +1,7 @@
+/** Issue Action Enums */
+export enum GithubAction {
+  Opened = 'opened',
+  Closed = 'closed',
+  Edited = 'edited',
+  Deleted = 'deleted'
+}
